+++ front/src/components/Navbar/Navbar.tsx
@@ -1,3 +1,4 @@
+"use client";
 import Link from "next/link";
 import React from "react";
 import logoFinal from "../../../public/images/logoFinal.png";
@@ -5,33 +6,33 @@ import Image from "next/image";
 
 const Navbar = () => {
   return (
-      <nav className="flex shadow-lg rounded-lg justify-between md:h-20 border border-[#3c9083]">
-        <div className="flex w-full justify-between">
-          <Link href="/">
-            <Image
-              src={logoFinal}
-              alt="logo"
-              width={60}
-              height={60}
-              className="ml-8"
-            ></Image>
-          </Link>
-          <div className="hidden w-3/5 items-center justify-evenly font-semibold md:flex">
-            <a href="/lostFound">Buscados / Encontrados</a>
-            <a href="not-found">Ayudanos donando</a>
-            <a href="/aboutUs">Sobre Huellas Unidas</a>
-            <a href="not-found">🐾Recomendaciones</a>
+    <nav className="flex shadow-lg rounded-lg justify-between md:h-20 border border-[#3c9083]">
+      <div className="flex w-full justify-between">
+        <Link href="/">
+          <Image
+            src={logoFinal}
+            alt="logo"
+            width={60}
+            height={60}
+            className="ml-8"
+          ></Image>
+        </Link>
+        <div className="hidden w-3/5 items-center justify-evenly font-semibold md:flex">
+          <Link href="/lostandfound">Buscados / Encontrados</Link>
+          <Link href="not-found">Ayudanos donando</Link>
+          <Link href="/aboutus">Sobre Huellas Unidas</Link>
+          <Link href="not-found">🐾Recomendaciones</Link>
+        </div>
+        <div className="hidden w-1/5 items-center justify-evenly font-semibold md:flex">
+          <div>
+            <Link href={"/register"}>Registrarme</Link>
           </div>
-          <div className="hidden w-1/5 items-center justify-evenly font-semibold md:flex">
-            <button>
-              <Link href={"/register"}>Registrarme</Link>
-            </button>
-            <button>
-              <Link href={"/login"}>Iniciar Sesión</Link>
-            </button>
+          <div>
+            <Link href={"/login"}>Iniciar Sesión</Link>
           </div>
         </div>
-      </nav>
+      </div>
+    </nav>
   );
 };
